refactor: split index.ts into 9 focused modules, scaffold Vitest
- config.ts: env var constants
- types.ts: all interfaces + Zod schemas (CategorySchema, TimeRangeSchema)
- domains.ts: domain filtering (loadDomainConfig, applyDomainFilters, etc.)
- cache.ts: Valkey client, cacheGet/Set/Clear, key generators
- reranker.ts: rerank(), rerankWithFallback()
- ollama.ts: expandQuery(), summarizePages(), formatSummaryResult()
- fetch.ts: assertPublicUrl(), fetchPage(), all fetch backends
- search.ts: searxSearchSingle(), searxSearch()
- tools.ts: registerTools() wrapping all 5 tool handlers
- index.ts: 14-line thin entry point

Also fixes a pre-existing SSRF gap: assertPublicUrl() now strips IPv6
brackets before pattern matching so [::1], [fc00::], [fe80::] addresses
are correctly blocked (Node.js URL.hostname preserves brackets for IPv6).

Adds Vitest 4.1.3 with tsconfig.test.json (bundler moduleResolution for
esbuild compat). 37 tests across domains, cache, fetch, and search:
- domains: urlMatchesDomain patterns, applyDomainFilters
- cache: searchCacheKey / fetchCacheKey consistency
- fetch: assertPublicUrl SSRF guard (13 cases including IPv6)
- search: CategorySchema / TimeRangeSchema Zod validation
