+++ package.json
@@ -5,7 +5,9 @@
   "main": "build/src/index.js",
   "type": "module",
   "scripts": {
-    "build": "tsc && chmod 755 build/src/index.js"
+    "build": "tsc && chmod 755 build/src/index.js",
+    "test": "vitest run --typecheck",
+    "test:watch": "vitest --typecheck"
   },
   "keywords": [
     "searxng",
@@ -21,6 +23,7 @@
   },
   "devDependencies": {
     "@types/node": "^22.13.10",
-    "typescript": "^5.8.2"
+    "typescript": "^5.8.2",
+    "vitest": "^4.1.3"
   }
 }
