+++ vitest.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    environment: "node",
+    typecheck: {
+      tsconfig: "./tsconfig.test.json",
+    },
+  },
+});
